requirements.txt
streamlit
torch
Pillow
torchvision
numpy
opencv-python
pyyaml
requests
